[TestInfra] 컨테이너 인스턴스 상수 제거 및 TestContainer 위치 이동
[refactor] legacy TestContainer에서 컨테이너 인스턴스 상수 및 헬퍼 메서드 제거
- MARIA_DB, REDIS 정적 인스턴스 상수 제거 — legacy 클래스들이 직접 컨테이너 생성하도록 변경
- getAndStart*(), registerMariaDb(), registerRedis() 헬퍼 메서드 제거
- legacy 3개 클래스(TestIntegrationContainer, V2, TestPersistenceContainer)가 각자 컨테이너를 직접 구성

[chore] TestContainer 상수 파일을 legacy → constant 패키지로 이동
- 이미지명, DB명, 포트 등 순수 상수만 남긴 TestContainer를 support.constant 패키지로 이전
- TestMariaDBContainer, TestRedisContainer, TestDataFactory import 경로 일괄 수정
